[UPDATE] XYZTupleをseichiassist.dataに移動する
