+++ src/main/scala/com/github/unchama/seichiassist/ActiveSkill.java
@@ -1,6 +1,6 @@
 package com.github.unchama.seichiassist;
 
-import com.github.unchama.seichiassist.effect.XYZTuple;
+import com.github.unchama.seichiassist.data.XYZTuple;
 import org.bukkit.Material;
 import org.bukkit.potion.PotionType;
