+++ src/main/scala/com/github/unchama/seichiassist/ActiveSkillPremiumEffect.scala
@@ -1,7 +1,7 @@
 package com.github.unchama.seichiassist
 
 import com.github.unchama.seichiassist.ActiveSkillPremiumEffect.plugin
-import com.github.unchama.seichiassist.effect.XYZTuple
+import com.github.unchama.seichiassist.data.XYZTuple
 import com.github.unchama.seichiassist.effect.arrow.ArrowEffects
 import com.github.unchama.seichiassist.effect.breaking.MagicTask
 import com.github.unchama.targetedeffect.TargetedEffect
