+++ src/main/scala/com/github/unchama/seichiassist/data/BreakArea.java
@@ -2,7 +2,7 @@ package com.github.unchama.seichiassist.data;
 
 
 import com.github.unchama.seichiassist.ActiveSkill;
-import com.github.unchama.seichiassist.effect.XYZTuple;
+import com.github.unchama.seichiassist.data.XYZTuple;
 import com.github.unchama.seichiassist.util.BreakUtil;
 import org.bukkit.entity.Player;
