+++ src/main/scala/com/github/unchama/seichiassist/data/syntax/package.scala
@@ -1,0 +1,5 @@
+package com.github.unchama.seichiassist.data
+
+package object syntax
+  extends AxisAlignedCuboidSyntax
+    with XYZTupleSyntax
