+++ src/main/scala/com/github/unchama/seichiassist/effect/breaking/BlizzardTask.scala
@@ -1,8 +1,6 @@
 package com.github.unchama.seichiassist.effect.breaking
 
 import com.github.unchama.seichiassist.data.ActiveSkillData
-import com.github.unchama.seichiassist.effect.XYZTuple
-import com.github.unchama.seichiassist.effect.XYZTuple.AxisAlignedCuboid
 import com.github.unchama.seichiassist.util.BreakUtil
 import com.github.unchama.seichiassist.{ActiveSkill, SeichiAssist}
 import org.bukkit.block.Block
