+++ src/main/scala/com/github/unchama/seichiassist/listener/EntityListener.java
@@ -4,7 +4,7 @@ import com.github.unchama.seichiassist.*;
 import com.github.unchama.seichiassist.data.BreakArea;
 import com.github.unchama.seichiassist.data.Mana;
 import com.github.unchama.seichiassist.data.player.PlayerData;
-import com.github.unchama.seichiassist.effect.XYZTuple;
+import com.github.unchama.seichiassist.data.XYZTuple;
 import com.github.unchama.seichiassist.task.GiganticBerserkTask;
 import com.github.unchama.seichiassist.util.BreakUtil;
 import com.github.unchama.seichiassist.util.Util;
